Move RefEq to its own file.
